cleanuo

+++ source/blender/nodes/geometry/nodes/node_geo_realize_instances.cc
@@ -33,7 +33,7 @@ static void node_declare(NodeDeclarationBuilder &b)
 
 static void node_init(bNodeTree * /*tree*/, bNode *node)
 {
-  node->custom1 = int(true);
+  node->custom1 = int16_t(true);
 }
 
 static void node_layout_ex(uiLayout *layout, bContext * /*C*/, PointerRNA *ptr)
